requirements.txt
absl-py==2.1.0
ansicon==1.89.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
blessed==1.20.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-aliases==1.0.4
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Django==5.1
executing==2.2.0
fastjsonschema==2.21.1
flatbuffers==24.3.25
flomodoro==1.1.0
fonttools==4.54.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.2
jinxed==1.3.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupytext==1.17.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mediapipe==0.10.18
ml-dtypes==0.4.1
MouseInfo==0.1.3
mysql-connector-python==9.4.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.8
playsound==1.2.2
plyer==2.1.0
prompt_toolkit==3.0.51
protobuf==4.25.5
psutil==7.0.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycparser==2.22
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pymunk==6.8.1
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rich==13.8.0
rpds-py==0.25.1
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.6.0
six==1.16.0
sounddevice==0.5.1
sqlparse==0.5.1
stack-data==0.6.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uv==0.6.3
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
